start simplification
