+++ slicer_cli_web/rest_slicer_cli.py
@@ -1,13 +1,10 @@
 import os
 import json
 import six
-import subprocess
 import sys
-import tempfile
 
-from ctk_cli import CLIModule
 from girder.api.rest import Resource, loadmodel, boundHandler, \
-    setResponseHeader, setRawResponse
+    RestException, setResponseHeader, setRawResponse
 from girder.api import access
 from girder.api.describe import Description, describeRoute
 from girder.constants import AccessType
@@ -15,6 +12,8 @@ from girder_worker.girder_plugin import utils as wutils
 from girder.utility.model_importer import ModelImporter
 from girder_worker.girder_plugin import constants
 from girder import logger
+
+from .cli_utils import as_model, generate_description
 
 _SLICER_TO_GIRDER_WORKER_TYPE_MAP = {
     'boolean': 'boolean',
@@ -92,36 +91,13 @@ def _getCLIParameters(clim):
     return index_params, opt_params, simple_out_params
 
 
-def _createIndexedParamTaskSpec(param):
-    """Creates task spec for indexed parameters
-
-    Parameters
-    ----------
-    param : ctk_cli.CLIParameter
-        parameter for which the task spec should be created
-
-    """
-
-    curTaskSpec = dict()
-    curTaskSpec['id'] = param.identifier()
-    curTaskSpec['name'] = param.label
-    curTaskSpec['type'] = _SLICER_TO_GIRDER_WORKER_TYPE_MAP[param.typ]
-    curTaskSpec['format'] = _SLICER_TO_GIRDER_WORKER_TYPE_MAP[param.typ]
-
-    if param.isExternalType():
-        curTaskSpec['target'] = 'filepath'  # check
-
-    return curTaskSpec
-
-
 def _addIndexedInputParamsToHandler(index_input_params, handlerDesc):
 
     for param in index_input_params:
 
         # add to route description
         if param.isExternalType():
-            suffix = _SLICER_TYPE_TO_GIRDER_INPUT_SUFFIX_MAP[param.typ]
-            handlerDesc.param(param.identifier() + suffix,
+            handlerDesc.param(param.identifier(),
                               'Girder ID of input %s - %s: %s'
                               % (param.typ, param.identifier(), param.description),
                               dataType='string', required=True)
@@ -130,43 +106,22 @@ def _addIndexedInputParamsToHandler(index_input_params, handlerDesc):
                               dataType='string', required=True)
 
 
-def _addIndexedInputParamsToTaskSpec(index_input_params, taskSpec):
-
-    for param in index_input_params:
-
-        # add to task spec
-        curTaskSpec = _createIndexedParamTaskSpec(param)
-        taskSpec['inputs'].append(curTaskSpec)
-
-
 def _addIndexedOutputParamsToHandler(index_output_params, handlerDesc):
 
     for param in index_output_params:
 
         # add param for parent folder to route description
-        handlerDesc.param(param.identifier() + _girderOutputFolderSuffix,
+        handlerDesc.param(param.identifier() + '_folder',
                           'Girder ID of parent folder '
                           'for output %s - %s: %s'
                           % (param.typ, param.typ, param.description),
                           dataType='string', required=True)
 
         # add param for name of current output to route description
-        handlerDesc.param(param.identifier() + _girderOutputNameSuffix,
+        handlerDesc.param(param.identifier(),
                           'Name of output %s - %s: %s'
                           % (param.typ, param.identifier(), param.description),
                           dataType='string', required=True)
-
-
-def _addIndexedOutputParamsToTaskSpec(index_output_params, taskSpec, hargs):
-
-    for param in index_output_params:
-
-        # add to task spec
-        curTaskSpec = _createIndexedParamTaskSpec(param)
-
-        curTaskSpec['path'] = hargs['params'][param.identifier() + _girderOutputNameSuffix]
-
-        taskSpec['outputs'].append(curTaskSpec)
 
 
 def _getParamDefaultVal(param):
@@ -185,36 +140,6 @@ def _getParamDefaultVal(param):
             'provide a default value in the xml' % param.typ)
 
 
-def _createOptionalParamTaskSpec(param):
-    """Creates task spec for optional parameters
-
-    Parameters
-    ----------
-    param : ctk_cli.CLIParameter
-        parameter for which the task spec should be created
-
-    """
-
-    curTaskSpec = dict()
-    curTaskSpec['id'] = param.identifier()
-    if getattr(param, 'label', None):
-        curTaskSpec['name'] = param.label
-    curTaskSpec['type'] = _SLICER_TO_GIRDER_WORKER_TYPE_MAP[param.typ]
-    curTaskSpec['format'] = _SLICER_TO_GIRDER_WORKER_TYPE_MAP[param.typ]
-
-    if param.isExternalType():
-        curTaskSpec['target'] = 'filepath'
-
-    if param.channel != 'output':
-
-        defaultValSpec = dict()
-        defaultValSpec['format'] = curTaskSpec['format']
-        defaultValSpec['data'] = _getParamDefaultVal(param)
-        curTaskSpec['default'] = defaultValSpec
-
-    return curTaskSpec
-
-
 def _addOptionalInputParamsToHandler(opt_input_params, handlerDesc):
 
     for param in opt_input_params:
@@ -223,8 +148,7 @@ def _addOptionalInputParamsToHandler(opt_input_params, handlerDesc):
         defaultVal = _getParamDefaultVal(param)
 
         if param.isExternalType():
-            suffix = _SLICER_TYPE_TO_GIRDER_INPUT_SUFFIX_MAP[param.typ]
-            handlerDesc.param(param.identifier() + suffix,
+            handlerDesc.param(param.identifier(),
                               'Girder ID of input %s - %s: %s'
                               % (param.typ, param.identifier(), param.description),
                               dataType='string',
@@ -235,16 +159,6 @@ def _addOptionalInputParamsToHandler(opt_input_params, handlerDesc):
                               default=json.dumps(defaultVal),
                               required=False)
 
-
-def _addOptionalInputParamsToTaskSpec(opt_input_params, taskSpec):
-
-    for param in opt_input_params:
-
-        # add to task spec
-        curTaskSpec = _createOptionalParamTaskSpec(param)
-        taskSpec['inputs'].append(curTaskSpec)
-
-
 def _addOptionalOutputParamsToHandler(opt_output_params, handlerDesc):
 
     for param in opt_output_params:
@@ -253,7 +167,7 @@ def _addOptionalOutputParamsToHandler(opt_output_params, handlerDesc):
             continue
 
         # add param for parent folder to route description
-        handlerDesc.param(param.identifier() + _girderOutputFolderSuffix,
+        handlerDesc.param(param.identifier() + '_folder',
                           'Girder ID of parent folder '
                           'for output %s - %s: %s'
                           % (param.typ, param.identifier(), param.description),
@@ -261,26 +175,10 @@ def _addOptionalOutputParamsToHandler(opt_output_params, handlerDesc):
                           required=False)
 
         # add param for name of current output to route description
-        handlerDesc.param(param.identifier() + _girderOutputNameSuffix,
+        handlerDesc.param(param.identifier(),
                           'Name of output %s - %s: %s'
                           % (param.typ, param.identifier(), param.description),
                           dataType='string', required=False)
-
-
-def _addOptionalOutputParamsToTaskSpec(opt_output_params, taskSpec, hargs):
-
-    for param in opt_output_params:
-        # set path if it was requested in the REST request
-        if (param.identifier() + _girderOutputFolderSuffix not in hargs['params'] or
-                param.identifier() + _girderOutputNameSuffix not in hargs['params']):
-            continue
-
-        # add to task spec
-        curTaskSpec = _createOptionalParamTaskSpec(param)
-
-        curTaskSpec['path'] = hargs['params'][param.identifier() + _girderOutputNameSuffix]
-
-        taskSpec['outputs'].append(curTaskSpec)
 
 
 def _addReturnParameterFileParamToHandler(handlerDesc):
@@ -290,7 +188,7 @@ def _addReturnParameterFileParamToHandler(handlerDesc):
     curDesc = _return_parameter_file_desc
 
     # add param for parent folder to route description
-    handlerDesc.param(curName + _girderOutputFolderSuffix,
+    handlerDesc.param(curName + '_folder',
                       'Girder ID of parent folder '
                       'for output %s - %s: %s'
                       % (curType, curName, curDesc),
@@ -298,286 +196,13 @@ def _addReturnParameterFileParamToHandler(handlerDesc):
                       required=False)
 
     # add param for name of current output to route description
-    handlerDesc.param(curName + _girderOutputNameSuffix,
+    handlerDesc.param(curName,
                       'Name of output %s - %s: %s'
                       % (curType, curName, curDesc),
                       dataType='string', required=False)
 
-
-def _addReturnParameterFileParamToTaskSpec(taskSpec, hargs):
-
-    curName = _return_parameter_file_name
-    curType = 'file'
-
-    # check if return parameter file was requested in the REST request
-    if (curName + _girderOutputFolderSuffix not in hargs['params'] or
-            curName + _girderOutputNameSuffix not in hargs['params']):
-        return
-
-    # add to task spec
-    curTaskSpec = dict()
-    curTaskSpec['id'] = curName
-    curTaskSpec['type'] = _SLICER_TO_GIRDER_WORKER_TYPE_MAP[curType]
-    curTaskSpec['format'] = _SLICER_TO_GIRDER_WORKER_TYPE_MAP[curType]
-    curTaskSpec['target'] = 'filepath'  # check
-    curTaskSpec['path'] = hargs['params'][curName + _girderOutputNameSuffix]
-
-    taskSpec['outputs'].append(curTaskSpec)
-
-
-def _createInputParamBindingSpec(param, hargs, token):
-
-    curBindingSpec = dict()
-    if _is_on_girder(param):
-        curBindingSpec = wutils.girderInputSpec(
-            hargs[param.identifier()],
-            resourceType=_SLICER_TYPE_TO_GIRDER_MODEL_MAP[param.typ],
-            dataType='string', dataFormat='string',
-            token=token, fetchParent=False)
-    else:
-        # inputs that are not of type image, file, or directory
-        # should be passed inline as string from json.dumps()
-        curBindingSpec['mode'] = 'inline'
-        curBindingSpec['type'] = _SLICER_TO_GIRDER_WORKER_TYPE_MAP[param.typ]
-        curBindingSpec['format'] = 'json'
-        curBindingSpec['data'] = hargs['params'][param.identifier()]
-
-    return curBindingSpec
-
-
-def _createOutputParamBindingSpec(param, hargs, user, token):
-
-    curBindingSpec = wutils.girderOutputSpec(
-        hargs[param.identifier()],
-        token,
-        name=hargs['params'][param.identifier() + _girderOutputNameSuffix],
-        dataType='string', dataFormat='string'
-    )
-
-    if param.isExternalType() and param.reference is not None:
-
-        if param.reference not in hargs:
-            raise Exception(
-                'Error: The specified reference attribute value'
-                '%s for parameter %s is not a valid input' % (
-                    param.reference, param.identifier())
-            )
-
-        curBindingSpec['reference'] = json.dumps({
-            'itemId': str(hargs[param.reference]['_id']),
-            'userId': str(user['_id']),
-            'identifier': param.identifier()
-        })
-
-    return curBindingSpec
-
-
-def _addIndexedInputParamBindings(index_input_params, bspec, hargs, token):
-
-    for param in index_input_params:
-        bspec[param.identifier()] = _createInputParamBindingSpec(param, hargs, token)
-
-
-def _addIndexedOutputParamBindings(index_output_params,
-                                   bspec, hargs, user, token):
-
-    for param in index_output_params:
-        bspec[param.identifier()] = _createOutputParamBindingSpec(
-            param, hargs, user, token)
-
-
-def _addOptionalInputParamBindings(opt_input_params, bspec, hargs, user, token):
-
-    for param in opt_input_params:
-
-        if _is_on_girder(param):
-            suffix = _SLICER_TYPE_TO_GIRDER_INPUT_SUFFIX_MAP[param.typ]
-            if param.identifier() + suffix not in hargs['params']:
-                continue
-
-            curModelName = _SLICER_TYPE_TO_GIRDER_MODEL_MAP[param.typ]
-            curModel = ModelImporter.model(curModelName)
-            curId = hargs['params'][param.identifier() + suffix]
-
-            hargs[param.identifier()] = curModel.load(id=curId,
-                                                      level=AccessType.READ,
-                                                      user=user)
-        if param.identifier() in hargs['params']:
-            bspec[param.identifier()] = _createInputParamBindingSpec(param, hargs, token)
-
-
-def _addOptionalOutputParamBindings(opt_output_params,
-                                    bspec, hargs, user, token):
-
-    for param in opt_output_params:
-
-        if not _is_on_girder(param):
-            continue
-
-        # check if it was requested in the REST request
-        if (param.identifier() + _girderOutputFolderSuffix not in hargs['params'] or
-                param.identifier() + _girderOutputNameSuffix not in hargs['params']):
-            continue
-
-        curModel = ModelImporter.model('folder')
-        curId = hargs['params'][param.identifier() + _girderOutputFolderSuffix]
-        doc = curModel.load(id=curId, level=AccessType.WRITE, user=user)
-        if doc:
-            hargs[param.identifier()] = doc
-
-        if param.identifier() in hargs:
-            bspec[param.identifier()] = _createOutputParamBindingSpec(param, hargs, user, token)
-
-
-def _addReturnParameterFileBinding(bspec, hargs, user, token, job):
-
-    curName = _return_parameter_file_name
-
-    # check if return parameter file was requested in the REST request
-    if (curName + _girderOutputFolderSuffix not in hargs['params'] or
-            curName + _girderOutputNameSuffix not in hargs['params']):
-        return
-
-    curModel = ModelImporter.model('folder')
-    curId = hargs['params'][curName + _girderOutputFolderSuffix]
-
-    hargs[curName] = curModel.load(id=curId,
-                                   level=AccessType.WRITE,
-                                   user=user)
-
-    curBindingSpec = wutils.girderOutputSpec(
-        hargs[curName],
-        token,
-        name=hargs['params'][curName + _girderOutputNameSuffix],
-        dataType='string', dataFormat='string',
-        reference=json.dumps({
-            'type': 'slicer_cli.parameteroutput',
-            'jobId': str(job['_id'])
-        })
-    )
-
-    bspec[curName] = curBindingSpec
-
-
 def _is_on_girder(param):
     return param.typ in _SLICER_TYPE_TO_GIRDER_MODEL_MAP
-
-
-def _getParamCommandLineValue(param, value):
-    if isinstance(value, six.binary_type):
-        value = value.decode('utf8')
-    if param.isVector():
-        cmdVal = '%s' % ', '.join(map(str, json.loads(value)))
-    else:
-        cmdVal = str(json.loads(value))
-
-    return cmdVal
-
-
-def _addOptionalInputParamsToContainerArgs(opt_input_params,
-                                           containerArgs, hargs):
-
-    for param in opt_input_params:
-
-        if param.longflag:
-            curFlag = param.longflag
-        elif param.flag:
-            curFlag = param.flag
-        else:
-            continue
-
-        if _is_on_girder(param) and param.identifier() in hargs:
-
-            curValue = "$input{%s}" % param.identifier()
-
-        elif param.identifier() in hargs['params']:
-
-            try:
-                curValue = _getParamCommandLineValue(
-                    param, hargs['params'][param.identifier()])
-            except Exception:
-                logger.exception(
-                    'Error: Parameter value is not in json.dumps format\n'
-                    '  Parameter name = %r\n  Parameter type = %r\n'
-                    '  Value passed = %r', param.identifier(), param.typ,
-                    hargs['params'][param.identifier()])
-                raise
-        else:
-            continue
-
-        containerArgs.append(curFlag)
-        containerArgs.append(curValue)
-
-
-def _addOptionalOutputParamsToContainerArgs(opt_output_params,
-                                            containerArgs, kwargs, hargs):
-    for param in opt_output_params:
-
-        if param.longflag:
-            curFlag = param.longflag
-        elif param.flag:
-            curFlag = param.flag
-        else:
-            continue
-        if _is_on_girder(param) and param.identifier() in kwargs['outputs']:
-
-            curValue = os.path.join(
-                _worker_docker_data_dir,
-                hargs['params'][param.identifier() + _girderOutputNameSuffix]
-            )
-
-            containerArgs.append(curFlag)
-            containerArgs.append(curValue)
-
-
-def _addReturnParameterFileToContainerArgs(containerArgs, kwargs, hargs):
-
-    curName = _return_parameter_file_name
-
-    if curName in kwargs['outputs']:
-
-        curFlag = '--returnparameterfile'
-
-        curValue = os.path.join(
-            _worker_docker_data_dir,
-            hargs['params'][curName + _girderOutputNameSuffix]
-        )
-
-        containerArgs.append(curFlag)
-        containerArgs.append(curValue)
-
-
-def _addIndexedParamsToContainerArgs(index_params, containerArgs, hargs):
-
-    for param in index_params:
-
-        if param.channel != 'output':
-
-            if _is_on_girder(param):
-                curValue = "$input{%s}" % param.identifier()
-            else:
-                curValue = _getParamCommandLineValue(
-                    param,
-                    hargs['params'][param.identifier()]
-                )
-
-        else:
-
-            if not _is_on_girder(param):
-                raise Exception(
-                    'The type of indexed output parameter %d '
-                    'must be of type - %s' % (
-                        param.index,
-                        _SLICER_TYPE_TO_GIRDER_MODEL_MAP.keys()
-                    )
-                )
-
-            curValue = os.path.join(
-                _worker_docker_data_dir,
-                hargs['params'][param.identifier() + _girderOutputNameSuffix]
-            )
-
-        containerArgs.append(curValue)
 
 
 def genHandlerToRunDockerCLI(dockerImage, cliRelPath, cliXML, restResource):
@@ -608,42 +233,23 @@ def genHandlerToRunDockerCLI(dockerImage, cliRelPath, cliXML, restResource):
     # get xml spec
     str_xml = cliXML
 
-    # parse cli xml spec
-    with tempfile.NamedTemporaryFile(suffix='.xml') as f:
-        f.write(str_xml if isinstance(str_xml, six.binary_type) else str_xml.encode('utf8'))
-        f.flush()
-        clim = CLIModule(f.name)
-
-    # create CLI description string
-    str_description = ['Description: <br/><br/>' + clim.description]
-
-    if clim.version is not None and len(clim.version) > 0:
-        str_description.append('Version: ' + clim.version)
-
-    if clim.license is not None and len(clim.license) > 0:
-        str_description.append('License: ' + clim.license)
-
-    if clim.contributor is not None and len(clim.contributor) > 0:
-        str_description.append('Author(s): ' + clim.contributor)
-
-    if clim.acknowledgements is not None and len(clim.acknowledgements) > 0:
-        str_description.append('Acknowledgements: ' + clim.acknowledgements)
-
-    str_description = '<br/><br/>'.join(str_description)
+    clim = as_model(str_xml)
 
     # do stuff needed to create REST endpoint for cLI
-    handlerDesc = Description(clim.title).notes(str_description)
+    handlerDesc = Description(clim.title).notes(generate_description(clim))
 
     # get CLI parameters
     index_params, opt_params, simple_out_params = _getCLIParameters(clim)
 
     # add indexed input parameters
     index_input_params = [p for p in index_params if p.channel != 'output']
+    index_input_params_on_girder = list(filter(_is_on_girder, index_input_params))
 
     _addIndexedInputParamsToHandler(index_input_params, handlerDesc)
 
     # add indexed output parameters
     index_output_params = [p for p in index_params if p.channel == 'output']
+    index_output_params_on_girder = list(filter(_is_on_girder, index_output_params))
 
     _addIndexedOutputParamsToHandler(index_output_params, handlerDesc)
 
@@ -667,118 +273,51 @@ def genHandlerToRunDockerCLI(dockerImage, cliRelPath, cliXML, restResource):
     @describeRoute(handlerDesc)
     def cliHandler(self, **hargs):
 
-        user = self.getCurrentUser()
-        token = self.getCurrentToken()['_id']
+        # verify that they are valid objects references
+        for param in index_input_params_on_girder:
+            curModel = ModelImporter.model(_SLICER_TYPE_TO_GIRDER_MODEL_MAP[param.typ])
+            curId = hargs[param.identifier()]
+            loaded = curModel.load(curId, level=AccessType.READ)
+            if not loaded:
+                raise RestException('Invalid %s id (%s).' % (curModel.name, str(curId)))
+
+        folderModel = ModelImporter.model('folder')
+        for param in index_output_params_on_girder:
+            curId = hargs[param.identifier() + '_folder']
+            loaded = folderModel.load(curId, level=AccessType.WRITE)
+            if not loaded:
+                raise RestException('Invalid Folder id (%s).' % (curModel.name, str(curId)))
 
         # create job
-        jobModel = self.model('job', 'jobs')
+        jobModel = ModelImporter.model('job', 'jobs')
         jobTitle = '.'.join((restResource.resourceName, cliName))
         job = jobModel.createJob(title=jobTitle,
                                  type=jobTitle,
                                  handler='worker_handler',
-                                 user=user)
-        kwargs = {
-            'validate': False,
-            'auto_convert': True,
-            'cleanup': True,
-            'inputs': dict(),
-            'outputs': dict()
-        }
-
+                                 user=self.getCurrentUser())
         # create job info
-        jobToken = jobModel.createJobToken(job)
-        kwargs['jobInfo'] = wutils.jobInfoSpec(job, jobToken)
+        job['token'] = jobModel.createJobToken(job)
+        job['celeryTaskName'] = 'slicer_cli_web.image_worker_tasks.run'
 
-        # initialize task spec
-        taskSpec = {'name': cliName,
-                    'mode': 'docker',
-                    'docker_image': dockerImage,
-                    'pull_image': False,
-                    'inputs': [],
-                    'outputs': []}
-
-        _addIndexedInputParamsToTaskSpec(index_input_params, taskSpec)
-
-        _addIndexedOutputParamsToTaskSpec(index_output_params, taskSpec, hargs)
-
-        _addOptionalInputParamsToTaskSpec(opt_input_params, taskSpec)
-
-        _addOptionalOutputParamsToTaskSpec(opt_output_params, taskSpec, hargs)
-
-        if len(simple_out_params) > 0:
-            _addReturnParameterFileParamToTaskSpec(taskSpec, hargs)
-
-        kwargs['task'] = taskSpec
-
-        # add input/output parameter bindings
-        _addIndexedInputParamBindings(index_input_params,
-                                      kwargs['inputs'], hargs, token)
-
-        _addIndexedOutputParamBindings(index_output_params,
-                                       kwargs['outputs'], hargs, user, token)
-
-        _addOptionalInputParamBindings(opt_input_params,
-                                       kwargs['inputs'], hargs, user, token)
-
-        _addOptionalOutputParamBindings(opt_output_params,
-                                        kwargs['outputs'], hargs, user, token)
-
-        if len(simple_out_params) > 0:
-            _addReturnParameterFileBinding(kwargs['outputs'],
-                                           hargs, user, token, job)
-
-        # construct container arguments
-        containerArgs = [cliRelPath]
-
-        _addOptionalInputParamsToContainerArgs(opt_input_params,
-                                               containerArgs, hargs)
-
-        _addOptionalOutputParamsToContainerArgs(opt_output_params,
-                                                containerArgs, kwargs, hargs)
-
-        _addReturnParameterFileToContainerArgs(containerArgs, kwargs, hargs)
-
-        _addIndexedParamsToContainerArgs(index_params,
-                                         containerArgs, hargs)
-
-        taskSpec['container_args'] = containerArgs
+        kwargs = {
+            'inputs': dict(),
+            'outputs': dict(),
+            'cliXML': cliXML,
+            'dockerImage': dockerImage,
+            'cliRelPath': cliRelPath,
+            'hargs': hargs
+        }
 
         # schedule job
         job['kwargs'] = kwargs
+
         job = jobModel.save(job)
         jobModel.scheduleJob(job)
 
         # return result
-        return jobModel.filter(job, user)
+        return jobModel.filter(job, self.getCurrentUser())
 
-    handlerFunc = cliHandler
-
-    # loadmodel stuff for indexed input params on girder
-    index_input_params_on_girder = list(filter(_is_on_girder, index_input_params))
-
-    for param in index_input_params_on_girder:
-
-        curModel = _SLICER_TYPE_TO_GIRDER_MODEL_MAP[param.typ]
-        suffix = _SLICER_TYPE_TO_GIRDER_INPUT_SUFFIX_MAP[param.typ]
-        curMap = {param.identifier() + suffix: param.identifier()}
-
-        handlerFunc = loadmodel(map=curMap,
-                                model=curModel,
-                                level=AccessType.READ)(handlerFunc)
-
-    # loadmodel stuff for indexed output params on girder
-    index_output_params_on_girder = list(filter(_is_on_girder, index_output_params))
-
-    for param in index_output_params_on_girder:
-
-        curModel = 'folder'
-        curMap = {param.identifier() + _girderOutputFolderSuffix: param.identifier()}
-
-        handlerFunc = loadmodel(map=curMap,
-                                model=curModel,
-                                level=AccessType.WRITE)(handlerFunc)
-
-    return handlerFunc
+    return cliHandler
 
 
 def genHandlerToGetDockerCLIXmlSpec(cliRelPath, cliXML, restResource):
@@ -822,136 +361,6 @@ def genHandlerToGetDockerCLIXmlSpec(cliRelPath, cliXML, restResource):
     return getXMLSpecHandler
 
 
-def genRESTEndPointsForSlicerCLIsInDocker(info, restResource, dockerImages):
-    """Generates REST end points for slicer CLIs placed in subdirectories of a
-    given root directory and attaches them to a REST resource with the given
-    name.
-
-    For each CLI, it creates:
-    * a GET Route (<apiURL>/`restResourceName`/<cliRelativePath>/xmlspec)
-    that returns the xml spec of the CLI
-    * a POST Route (<apiURL>/`restResourceName`/<cliRelativePath>/run)
-    that runs the CLI
-
-    It also creates a GET route (<apiURL>/`restResourceName`) that returns a
-    list of relative routes to all CLIs attached to the generated REST resource
-
-    Parameters
-    ----------
-    info
-    restResource : str or girder.api.rest.Resource
-        REST resource to which the end-points should be attached
-    dockerImages : a list of docker image names
-
-    """
-    dockerImages
-    # validate restResource argument
-    if not isinstance(restResource, (str, Resource)):
-        raise Exception('restResource must either be a string or '
-                        'an object of girder.api.rest.Resource')
-
-    # validate dockerImages arguments
-    if not isinstance(dockerImages, (str, list)):
-        raise Exception('dockerImages must either be a single docker image '
-                        'string or a list of docker image strings')
-
-    if isinstance(dockerImages, list):
-        if any(not isinstance(img, str) for img in dockerImages):
-            raise Exception(
-                'dockerImages must either be a single docker image string or '
-                'a list of docker image strings')
-    else:
-        dockerImages = [dockerImages]
-
-    # create REST resource if given a name
-    if isinstance(restResource, str):
-        restResource = type(restResource,
-                            (Resource, ),
-                            {'resourceName': restResource})()
-
-    restResourceName = type(restResource).__name__
-
-    # Add REST routes for slicer CLIs in each docker image
-    cliList = []
-
-    for dimg in dockerImages:
-        # check if the docker image exists
-
-        getDockerImage(dimg, True)
-
-        # get CLI list
-        cliListSpec = getDockerImageCLIList(dimg)
-
-        if isinstance(cliListSpec, six.binary_type):
-            cliListSpec = cliListSpec.decode('utf8')
-        cliListSpec = json.loads(cliListSpec)
-
-        # Add REST end-point for each CLI
-        for cliRelPath in cliListSpec.keys():
-            cliXML = getDockerImageCLIXMLSpec(dimg, cliRelPath)
-            # create a POST REST route that runs the CLI
-            try:
-
-                cliRunHandler = genHandlerToRunDockerCLI(dimg,
-                                                         cliRelPath,
-                                                         cliXML,
-                                                         restResource)
-            except Exception:
-                logger.execption('Failed to create REST endpoints for %s',
-                                 cliRelPath)
-                continue
-
-            cliSuffix = os.path.normpath(cliRelPath).replace(os.sep, '_')
-
-            cliRunHandlerName = 'run_' + cliSuffix
-            setattr(restResource, cliRunHandlerName, cliRunHandler)
-            restResource.route('POST',
-                               (cliRelPath, 'run'),
-                               getattr(restResource, cliRunHandlerName))
-
-            # create GET REST route that returns the xml of the CLI
-            try:
-                cliGetXMLSpecHandler = genHandlerToGetDockerCLIXmlSpec(
-                    cliRelPath, cliXML, restResource)
-
-            except Exception:
-                logger.exception('Failed to create REST endpoints for %s',
-                                 cliRelPath)
-                exc_type, exc_obj, exc_tb = sys.exc_info()
-                fname = os.path.split(exc_tb.tb_frame.f_code.co_filename)[1]
-                logger.error('%r' % [exc_type, fname, exc_tb.tb_lineno])
-                continue
-
-            cliGetXMLSpecHandlerName = 'get_xml_' + cliSuffix
-            setattr(restResource,
-                    cliGetXMLSpecHandlerName,
-                    cliGetXMLSpecHandler)
-            restResource.route('GET',
-                               (cliRelPath, 'xmlspec',),
-                               getattr(restResource, cliGetXMLSpecHandlerName))
-
-            cliList.append(cliRelPath)
-
-    # create GET route that returns a list of relative routes to all CLIs
-    @boundHandler(restResource)
-    @access.user
-    @describeRoute(
-        Description('Get list of relative routes to all CLIs')
-    )
-    def getCLIListHandler(self, *args, **kwargs):
-        return cliList
-
-    getCLIListHandlerName = 'get_cli_list'
-    setattr(restResource, getCLIListHandlerName, getCLIListHandler)
-    restResource.route('GET', (), getattr(restResource, getCLIListHandlerName))
-
-    # expose the generated REST resource via apiRoot
-    setattr(info['apiRoot'], restResourceName, restResource)
-
-    # return restResource
-    return restResource
-
-
 def genRESTEndPointsForSlicerCLIsInDockerCache(restResource, dockerCache):
     """Generates REST end points for slicer CLIs placed in subdirectories of a
     given root directory and attaches them to a REST resource with the given
@@ -974,21 +383,18 @@ def genRESTEndPointsForSlicerCLIsInDockerCache(restResource, dockerCache):
 
     """
 
-    dockerImages = dockerCache.getImageNames()
     # validate restResource argument
     if not isinstance(restResource, Resource):
         raise Exception('restResource must be a Docker Resource')
 
-    for dimg in dockerImages:
-
-        docker_image = dockerCache.getImageByName(dimg)
+    for docker_image in dockerCache.getImages():
+        dimg = docker_image.name
         # get CLI list
         cliListSpec = docker_image.getCLIListSpec()
 
+        restPath = dimg.replace(':', '_').replace('/', '_').replace('@', '_')
         # Add REST end-point for each CLI
         for cliRelPath in cliListSpec.keys():
-            restPath = dimg.replace(
-                ':', '_').replace('/', '_').replace('@', '_')
             # create a POST REST route that runs the CLI
             try:
                 cliXML = docker_image.getCLIXML(cliRelPath)
@@ -1040,83 +446,3 @@ def genRESTEndPointsForSlicerCLIsInDockerCache(restResource, dockerCache):
             logger.debug('Created REST endpoints for %s', cliRelPath)
 
     return restResource
-
-
-def getDockerImage(imageName, pullIfNotExist=False):
-    """
-    Checks the local docker cache for the image
-
-    :param imageName: the docker image name in the form of repo/name:tag
-    if the tag is not given docker defaults to using the :latest tag
-    :type imageName: string
-    :returns: if the image exists the id(sha256 hash) is returned otherwise
-    None is returned
-    """
-    try:
-        # docker inspect returns non zero if the image is not available
-        # locally
-        data = subprocess.check_output(['docker', 'inspect',
-                                        '--format="{{json .Id}}"', imageName])
-
-        return data
-    except subprocess.CalledProcessError:
-        if pullIfNotExist:
-            # the image does not exist locally, try to pull from dockerhub
-            # none is returned if it fails
-            data = pullDockerImage(imageName)
-            return data
-        raise Exception('cant find the image %s' % imageName)
-
-
-def getDockerImageCLIList(imageName):
-    """
-    Gets the cli list of the docker image.
-
-    :param imageName: the docker image name in the form of repo/name:tag
-    :type imageName: string
-    :returns: if the image exist the cli dictionary is returned otherwise
-    None is returned
-    cli dictionary format is the following:
-    {
-    <cli_name>:{
-                type:<type>
-                }
-
-    }
-    """
-    try:
-        # docker inspect returns non zero if the image is not available
-        # locally
-        data = subprocess.check_output(
-            ['docker', 'run', '--rm', imageName, '--list_cli'])
-        return data
-    except subprocess.CalledProcessError:
-        # the image does not exist locally, try to pull from dockerhub
-        raise Exception("Could not get the cli list for the img %s most "
-                        "likely the docker image entrypoint does "
-                        "not accept the argument --list_cli" % imageName)
-
-
-def getDockerImageCLIXMLSpec(img, cli):
-    """
-    Gets the xml spec of the specific cli
-
-    """
-    try:
-        data = subprocess.check_output(['docker', 'run', '--rm', img, cli, '--xml'])
-        return data
-    except subprocess.CalledProcessError:
-        # the image does not exist locally, try to pull from dockerhub
-        raise Exception('Could not get xml data for img %s '
-                        'cli %s' % (img, cli))
-
-
-def pullDockerImage(img):
-    try:
-        subprocess.check_output(['docker', 'pull', img])
-        data = getDockerImage(img)
-        return data
-    except subprocess.CalledProcessError:
-        # the image does not exist on the default repository
-
-        return None
